pyproject.toml
[project]
name = "aeroml"
version = "0.1.0"
description = "AERO: Automated Exploration, Research & Orchestration – A Framework for Machine Learning Research"
readme = "README.md"
authors = [
    { name = "Jacob Wong", email = "" },
    { name = "Ethan Lau", email = "" },
    { name = "Charmaine Chua", email = "chuacharmaine648@gmail.com" },
]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.9.0",
    "faiss-cpu>=1.7.0",
    "feedparser>=6.0.0",
    "h5py>=3.11.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.40",
    "litellm>=1.0.0",
    "matplotlib>=3.5.0",
    "numpy>=1.20.0",
    "openai>=1.0.0",
    "openpyxl>=3.1.0",
    "pandas>=1.3.0",
    "pypdf2>=3.0.0",
    "python-docx>=0.8.11",
    "python-dotenv>=0.19.0",
    "python-json-logger>=2.0.7",
    "requests>=2.25.0",
    "sentence-transformers>=2.0.0",
    "soupsieve>=2.8",
    "tavily-python>=0.3.0",
    "tiktoken>=0.5.0",
    "torch>=1.9.0",
    "transformers>=4.0.0",
    "treequest==0.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aero"]

[tool.hatch.build.targets.sdist]
packages = ["src/aero"]